+++ packages/dsi-logo-maker/functions/fillSVG/index.js
@@ -6,19 +6,31 @@ import {
   precomputeBlocks,
   getIndexModule,
 } from "../../utils/blocks";
+import { useLoadedOpentypeFont } from "../../utils/hooks";
 import { Block } from "../../utils/blocks-components";
 
 export const handler = ({ inputs, mechanic }) => {
-  const { width, height, logoWidth, logoRatio } = inputs;
+  const { width, height, logoWidth, logoRatio, fontMode } = inputs;
   const { done } = mechanic;
 
   const rows = 2;
   const cols = 13;
   const logoHeight = Math.floor((logoWidth / logoRatio) * rows);
   const words = ["DESIGN", "SYSTEMS", "INTERNATIONAL"];
+  const font = useLoadedOpentypeFont(fontMode);
+
+  useEffect(() => {
+    if (font) {
+      done();
+    }
+  }, [font]);
+
+  if (!font) {
+    return null;
+  }
 
   const blockGeometry = computeBlockGeometry(logoWidth, logoHeight, rows, cols);
-  const baseBricks = computeBaseBricks(words, blockGeometry.fontSize);
+  const baseBricks = computeBaseBricks(words, blockGeometry.fontSize, font);
   const blocksByIndex = precomputeBlocks(blockGeometry, baseBricks);
 
   const blockConfigs = [];
@@ -41,10 +53,6 @@ export const handler = ({ inputs, mechanic }) => {
     }
   }
 
-  useEffect(() => {
-    done();
-  }, []);
-
   return (
     <svg width={width} height={height}>
       {blockConfigs.map(({ position, block, colors }) => (
@@ -62,17 +70,17 @@ export const handler = ({ inputs, mechanic }) => {
 export const inputs = {
   width: {
     type: "number",
-    default: 300,
+    default: 500,
     min: 100,
   },
   height: {
     type: "number",
-    default: 300,
+    default: 500,
     min: 100,
   },
   logoWidth: {
     type: "number",
-    default: 80,
+    default: 300,
     min: 10,
   },
   logoRatio: {
@@ -82,6 +90,14 @@ export const inputs = {
     slider: true,
     min: 6,
     step: 1,
+  },
+  fontMode: {
+    type: "text",
+    options: {
+      "F Grotesk Thin": "FGroteskThin-Regular.otf",
+      "F Grotesk": "FGrotesk-Regular.otf",
+    },
+    default: "F Grotesk Thin",
   },
 };