+++ packages/dsi-logo-maker/functions/logoAnimatedCanvas/index.js
@@ -5,12 +5,14 @@ import {
   precomputeBlocks,
   getIndexModule,
 } from "../../utils/blocks";
+import { loadOpentypeFont } from "../../utils/opentype";
 import { drawBlock } from "../../utils/blocks-canvas";
 
-export const handler = ({ inputs, mechanic }) => {
+export const handler = async ({ inputs, mechanic }) => {
   const {
     width,
     ratio,
+    fontMode,
     colorMode,
     flag,
     firstColor,
@@ -24,8 +26,8 @@ export const handler = ({ inputs, mechanic }) => {
   const rows = 2;
   const cols = 13;
   const words = ["DESIGN", "SYSTEMS", "INTERNATIONAL"];
-
   const height = Math.floor((width / ratio) * rows);
+  const font = await loadOpentypeFont(fontMode);
 
   const colors = getColors(colorMode, flag, [
     firstColor,
@@ -33,7 +35,7 @@ export const handler = ({ inputs, mechanic }) => {
     thirdColor,
   ]);
   const blockGeometry = computeBlockGeometry(width, height, rows, cols);
-  const baseBricks = computeBaseBricks(words, blockGeometry.fontSize);
+  const baseBricks = computeBaseBricks(words, blockGeometry.fontSize, font);
   const blocksByIndex = precomputeBlocks(blockGeometry, baseBricks);
   const position = { x: 0, y: 0 };
 
@@ -95,6 +97,14 @@ export const inputs = {
     slider: true,
     min: 6,
     step: 1,
+  },
+  fontMode: {
+    type: "text",
+    options: {
+      "F Grotesk Thin": "FGroteskThin-Regular.otf",
+      "F Grotesk": "FGrotesk-Regular.otf",
+    },
+    default: "F Grotesk Thin",
   },
   colorMode: {
     type: "text",